fix networking support
We did not do a proper config downgrade in best effort mode for the
network access rights.  This shadowed another bug in the definition of
the PathBeneathAttr struct, whose port number was wrongly passed using
16 instead of 64 bits (probably carried over from an earlier version
of the kernel patch).

+++ landlock/syscall/landlock.go
@@ -75,5 +75,5 @@ type PathBeneathAttr struct {
 // NetPortAttr specifies which ports can be used for what.
 type NetPortAttr struct {
 	AllowedAccess uint64
-	Port          uint16
+	Port          uint64
 }
